backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-celery-beat>=2.6,<3.0
dj-database-url>=2.1,<3.0
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0
psycopg[binary]>=3.1,<4.0
openai>=1.30,<2.0
Jinja2>=3.1,<4.0
gunicorn>=22.0,<23.0
pgvector>=0.3,<1.0
reportlab>=4.1,<5.0
python-docx>=1.1,<2.0
